Update strings.f90
Update str_split.  Correcting logic that was causing crashes with Intel compilers.

The function will now append a delimiter to the end of the input string prior to work to get the column string.  This is done because the logic of finding the column is based on a right-justified delimiter.